separating ApiEvent component from App
