+++ package.json
@@ -29,5 +29,8 @@
     "concurrently": "^7.6.0",
     "cross-env": "^7.0.3",
     "husky": "^8.0.3"
+  },
+  "dependencies": {
+    "react-leaflet": "^4.2.1"
   }
 }
